Split classes/interface into separate files, add IConfigManager
