+++ osu.Framework/Configuration/IConfigManager.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+namespace osu.Framework.Configuration
+{
+    public interface IConfigManager
+    {
+        void Load();
+        bool Save();
+    }
+}
