+++ osu.Framework/Configuration/Tracking/ITrackedSettings.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+namespace osu.Framework.Configuration.Tracking
+{
+    public interface ITrackedSettings
+    {
+        
+    }
+}
